+++ src/app/services/app.service.ts
@@ -12,8 +12,12 @@ import { StorageService } from './storage.service';
 export class AppService {
 
   appName:string = 'delibcanvas'
+
   remoteFieldMappingURL:string = 'https://raw.githubusercontent.com/uppy01/deliberative-canvas-config/main/fieldmapping.json'
-  
+  //for development & testing only...
+  //remoteFieldMappingURL:string = 'https://raw.githubusercontent.com/uppy01/deliberative-canvas-config/main/fieldmapping-DEV.json'
+
+
   authConfigSubscription:Subscription
   //replica:Earthstar.Replica
   user:EarthstarAuthorAddress