+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+#### Version 0.5.0 (2025.01.21)
+
+Features
+- import a csv without using pre-defined field mappings (note: json imports always require field mappings)
+
+Bugs & Fixes
+- exporting a 'date' field should not include time - just YYYY-MM-DD
+- when importing csv, use source field name when field mappings do not have an export field name set
+- when importing csv, any headers/fields that are not defined in the field mappings should be excluded from the export
+
+
 #### Version 0.1.0 (2024.09.14)
 The first (alpha) release of the Deliberative Canvas app contains the following major features:
 - "adapter pipeline" for importing csv/json files from other deliberative tools, and transforming the data for display on a canvas